SupressionLoaderTest have been splitted to two parts. Online, which may be unstable and offline, which has to be stable all the time
