+++ src/test/java/com/puppycrawl/tools/checkstyle/filters/SuppressionsLoaderTest.java
@@ -22,12 +22,8 @@ package com.puppycrawl.tools.checkstyle.filters;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.regex.PatternSyntaxException;
 
-import org.junit.Assume;
 import org.junit.Test;
 
 import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
@@ -48,42 +44,6 @@ public class SuppressionsLoaderTest {
         assertEquals(fc, fc2);
     }
 
-    @Test
-    public void testLoadFromURL()
-        throws CheckstyleException, InterruptedException {
-        boolean online = isInternetReachable();
-
-        Assume.assumeTrue(online);
-
-        FilterSet fc = null;
-
-        int attemptCount = 0;
-        final int attemptLimit = 5;
-        while (attemptCount <= attemptLimit) {
-            try {
-
-                fc = SuppressionsLoader
-                        .loadSuppressions("http://checkstyle.sourceforge.net/files/suppressions_none.xml");
-                break;
-
-            }
-            catch (CheckstyleException ex) {
-                // for some reason Travis CI failed some times(unstable) on reading this file
-                if (attemptCount < attemptLimit
-                        && ex.getMessage().contains("unable to read")) {
-                    attemptCount++;
-                    // wait for bad/disconnection time to pass
-                    Thread.sleep(1000);
-                }
-                else {
-                    throw ex;
-                }
-            }
-        }
-
-        final FilterSet fc2 = new FilterSet();
-        assertEquals(fc, fc2);
-    }
 
     @Test(expected = CheckstyleException.class)
     public void testLoadFromMalformedURL() throws CheckstyleException {
@@ -160,16 +120,4 @@ public class SuppressionsLoaderTest {
         }
     }
 
-    private static boolean isInternetReachable() {
-        try {
-            URL url = new URL("http://www.yahoo.com");
-            HttpURLConnection urlConnect = (HttpURLConnection) url.openConnection();
-            @SuppressWarnings("unused")
-            Object objData = urlConnect.getContent();
-        }
-        catch (IOException e) {
-            return false;
-        }
-        return true;
-    }
 }
